feat: better token counter
